picam test server

+++ videoServer.py
@@ -304,7 +304,7 @@ def audio_callback(indata, frames, time, status):
     """Callback to process audio input."""
     global bark_detected
     global rms
-    rms = max(np.sqrt(np.mean(indata[:, 0]**2)), np.sqrt(np.mean(indata[:, 1]**2)))
+    rms = float(max(np.sqrt(np.mean(indata[:, 0]**2)), np.sqrt(np.mean(indata[:, 1]**2))))
     print(rms)
     if rms > LOUDNESS_THRESHOLD:
         bark_detected = True
